+++ mleap-tensorflow/src/test/scala/ml/combust/mleap/tensorflow/TestUtil.scala
@@ -1,12 +1,16 @@
 package ml.combust.mleap.tensorflow
 
 import java.io.File
-import java.nio.file.{Files, Path}
 
 import org.tensorflow
 import org.tensorflow.op.Ops
 import org.tensorflow.types.TFloat32
-
+import org.tensorflow.Signature
+import org.tensorflow.op.core.Placeholder
+import org.tensorflow.ConcreteFunction
+import org.tensorflow.ndarray.Shape
+import java.nio.file.Files
+import java.nio.file.Path
 /**
   * Created by hollinwilkins on 1/13/17.
   * Updated to TF2.0 by AustinZh on 6/3/2021
@@ -39,5 +43,18 @@ object TestUtil {
     val inputB = tf.withName("InputB").placeholder(classOf[TFloat32])
     tf.withName("MyResult").math.mul(inputA, inputB)
     graph
+  }
+
+
+  def createConcreteFunctionWithVariables(xShape: Shape) :  ConcreteFunction = {
+    val g = new tensorflow.Graph()
+    val tf = tensorflow.op.Ops.create(g)
+    val x = tf.placeholder(classOf[TFloat32], Placeholder.shape(xShape))
+    val y = tf.variable(tf.random.randomUniform(tf.constant(xShape), classOf[TFloat32]))
+    val z = tf.reduceSum(tf.math.add(x, y), tf.array(0, 1))
+    val signature= Signature.builder.input("input", x).output("reducedSum", z).build
+    val session = new tensorflow.Session(g)
+    session.run(tf.init())
+    tensorflow.ConcreteFunction.create(signature, session)
   }
 }