+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 
 ### New Features
 - MathBinaryModel now supports Logit operations
+- TensorflowTransformerOp now supports serialization and deserialization using SavedModel format
 - `Casting.cast` now supports conversions between ListType and TensorType
 
 ### Bug Fixes